Move non-SSL tests to derived class
