+++ streams/src/test/java/org/apache/kafka/streams/integration/AbstractResetIntegrationTest.java
@@ -52,18 +52,15 @@ import org.junit.rules.TemporaryFolder;
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileWriter;
-import java.text.SimpleDateFormat;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Calendar;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 
 import static java.time.Duration.ofMillis;
-import static org.apache.kafka.streams.integration.utils.IntegrationTestUtils.isEmptyConsumerGroup;
 import static org.apache.kafka.streams.integration.utils.IntegrationTestUtils.waitForEmptyConsumerGroup;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -74,8 +71,8 @@ public abstract class AbstractResetIntegrationTest {
     static EmbeddedKafkaCluster cluster;
 
     private static MockTime mockTime;
-    private static KafkaStreams streams;
-    private static Admin adminClient = null;
+    protected static KafkaStreams streams;
+    protected static Admin adminClient = null;
 
     abstract Map<String, Object> getClientSslConfig();
 
@@ -87,11 +84,11 @@ public abstract class AbstractResetIntegrationTest {
         }
     }
 
-    private String appID = "abstract-reset-integration-test";
-    private Properties commonClientConfig;
-    private Properties streamsConfig;
+    protected String appID = "abstract-reset-integration-test";
+    protected Properties commonClientConfig;
+    protected Properties streamsConfig;
     private Properties producerConfig;
-    private Properties resultConsumerConfig;
+    protected Properties resultConsumerConfig;
 
     private void prepareEnvironment() {
         if (adminClient == null) {
@@ -158,16 +155,15 @@ public abstract class AbstractResetIntegrationTest {
     @Rule
     public final TemporaryFolder testFolder = new TemporaryFolder(TestUtils.tempDirectory());
 
-    private static final String INPUT_TOPIC = "inputTopic";
-    private static final String OUTPUT_TOPIC = "outputTopic";
+    protected static final String INPUT_TOPIC = "inputTopic";
+    protected static final String OUTPUT_TOPIC = "outputTopic";
     private static final String OUTPUT_TOPIC_2 = "outputTopic2";
     private static final String OUTPUT_TOPIC_2_RERUN = "outputTopic2_rerun";
     private static final String INTERMEDIATE_USER_TOPIC = "userTopic";
-    private static final String NON_EXISTING_TOPIC = "nonExistingTopic";
 
-    private static final int STREAMS_CONSUMER_TIMEOUT = 2000;
-    private static final int CLEANUP_CONSUMER_TIMEOUT = 2000;
-    private static final int TIMEOUT_MULTIPLIER = 15;
+    protected static final int STREAMS_CONSUMER_TIMEOUT = 2000;
+    protected static final int CLEANUP_CONSUMER_TIMEOUT = 2000;
+    protected static final int TIMEOUT_MULTIPLIER = 15;
 
     void prepareTest() throws Exception {
         prepareConfigs();
@@ -204,98 +200,6 @@ public abstract class AbstractResetIntegrationTest {
             mockTime.sleep(10);
             IntegrationTestUtils.produceKeyValuesSynchronouslyWithTimestamp(INPUT_TOPIC, Collections.singleton(record), producerConfig, mockTime.milliseconds());
         }
-    }
- 
-    public void shouldNotAllowToResetWhileStreamsIsRunning() {
-        appID = testId + "-not-reset-during-runtime";
-        final String[] parameters = new String[] {
-            "--application-id", appID,
-            "--bootstrap-servers", cluster.bootstrapServers(),
-            "--input-topics", NON_EXISTING_TOPIC,
-            "--execute"
-        };
-        final Properties cleanUpConfig = new Properties();
-        cleanUpConfig.put(ConsumerConfig.HEARTBEAT_INTERVAL_MS_CONFIG, 100);
-        cleanUpConfig.put(ConsumerConfig.SESSION_TIMEOUT_MS_CONFIG, Integer.toString(CLEANUP_CONSUMER_TIMEOUT));
-
-        streamsConfig.put(StreamsConfig.APPLICATION_ID_CONFIG, appID);
-
-        // RUN
-        streams = new KafkaStreams(setupTopologyWithoutIntermediateUserTopic(), streamsConfig);
-        streams.start();
-
-        final int exitCode = new StreamsResetter().run(parameters, cleanUpConfig);
-        Assert.assertEquals(1, exitCode);
-
-        streams.close();
-    }
-
-    public void shouldNotAllowToResetWhenInputTopicAbsent() throws Exception {
-        appID = testId + "-not-reset-without-input-topic";
-        final String[] parameters = new String[] {
-            "--application-id", appID,
-            "--bootstrap-servers", cluster.bootstrapServers(),
-            "--input-topics", NON_EXISTING_TOPIC,
-            "--execute"
-        };
-        final Properties cleanUpConfig = new Properties();
-        cleanUpConfig.put(ConsumerConfig.HEARTBEAT_INTERVAL_MS_CONFIG, 100);
-        cleanUpConfig.put(ConsumerConfig.SESSION_TIMEOUT_MS_CONFIG, Integer.toString(CLEANUP_CONSUMER_TIMEOUT));
-
-        final int exitCode = new StreamsResetter().run(parameters, cleanUpConfig);
-        Assert.assertEquals(1, exitCode);
-    }
-
-    public void shouldNotAllowToResetWhenIntermediateTopicAbsent() throws Exception {
-        appID = testId + "-not-reset-without-intermediate-topic";
-        final String[] parameters = new String[] {
-            "--application-id", appID,
-            "--bootstrap-servers", cluster.bootstrapServers(),
-            "--intermediate-topics", NON_EXISTING_TOPIC,
-            "--execute"
-        };
-        final Properties cleanUpConfig = new Properties();
-        cleanUpConfig.put(ConsumerConfig.HEARTBEAT_INTERVAL_MS_CONFIG, 100);
-        cleanUpConfig.put(ConsumerConfig.SESSION_TIMEOUT_MS_CONFIG, Integer.toString(CLEANUP_CONSUMER_TIMEOUT));
-
-        final int exitCode = new StreamsResetter().run(parameters, cleanUpConfig);
-        Assert.assertEquals(1, exitCode);
-    }
-
-    @Test
-    public void testResetWhenLongSessionTimeoutConfiguredWithForceOption() throws Exception {
-        appID = testId + "-with-force-option";
-        streamsConfig.put(StreamsConfig.APPLICATION_ID_CONFIG, appID);
-        streamsConfig.put(ConsumerConfig.SESSION_TIMEOUT_MS_CONFIG, Integer.toString(STREAMS_CONSUMER_TIMEOUT * 100));
-
-        // Run
-        streams = new KafkaStreams(setupTopologyWithoutIntermediateUserTopic(), streamsConfig);
-        streams.start();
-        final List<KeyValue<Long, Long>> result = IntegrationTestUtils.waitUntilMinKeyValueRecordsReceived(resultConsumerConfig, OUTPUT_TOPIC, 10);
-
-        streams.close();
-
-        // RESET
-        streams = new KafkaStreams(setupTopologyWithoutIntermediateUserTopic(), streamsConfig);
-        streams.cleanUp();
-
-        // Reset would fail since long session timeout has been configured
-        final boolean cleanResult = tryCleanGlobal(false, null, null);
-        Assert.assertFalse(cleanResult);
-
-        // Reset will success with --force, it will force delete active members on broker side
-        cleanGlobal(false, "--force", null);
-        assertThat("Group is not empty after cleanGlobal", isEmptyConsumerGroup(adminClient, appID));
-
-        assertInternalTopicsGotDeleted(null);
-
-        // RE-RUN
-        streams.start();
-        final List<KeyValue<Long, Long>> resultRerun = IntegrationTestUtils.waitUntilMinKeyValueRecordsReceived(resultConsumerConfig, OUTPUT_TOPIC, 10);
-        streams.close();
-
-        assertThat(resultRerun, equalTo(result));
-        cleanGlobal(false, "--force", null);
     }
 
     @Test
@@ -340,7 +244,7 @@ public abstract class AbstractResetIntegrationTest {
         testReprocessingFromScratchAfterResetWithIntermediateUserTopic(true);
     }
 
-    public void testReprocessingFromScratchAfterResetWithIntermediateUserTopic(final boolean useRepartitioned) throws Exception {
+    private void testReprocessingFromScratchAfterResetWithIntermediateUserTopic(final boolean useRepartitioned) throws Exception {
         if (!useRepartitioned) {
             cluster.createTopic(INTERMEDIATE_USER_TOPIC);
         }
@@ -405,132 +309,6 @@ public abstract class AbstractResetIntegrationTest {
         }
     }
 
-    @Test
-    public void testReprocessingFromFileAfterResetWithoutIntermediateUserTopic() throws Exception {
-        appID = testId + "-from-file";
-        streamsConfig.put(StreamsConfig.APPLICATION_ID_CONFIG, appID);
-
-        // RUN
-        streams = new KafkaStreams(setupTopologyWithoutIntermediateUserTopic(), streamsConfig);
-        streams.start();
-        final List<KeyValue<Long, Long>> result = IntegrationTestUtils.waitUntilMinKeyValueRecordsReceived(resultConsumerConfig, OUTPUT_TOPIC, 10);
-
-        streams.close();
-        waitForEmptyConsumerGroup(adminClient, appID, TIMEOUT_MULTIPLIER * STREAMS_CONSUMER_TIMEOUT);
-
-        // RESET
-        final File resetFile = File.createTempFile("reset", ".csv");
-        try (final BufferedWriter writer = new BufferedWriter(new FileWriter(resetFile))) {
-            writer.write(INPUT_TOPIC + ",0,1");
-        }
-
-        streams = new KafkaStreams(setupTopologyWithoutIntermediateUserTopic(), streamsConfig);
-        streams.cleanUp();
-
-        cleanGlobal(false, "--from-file", resetFile.getAbsolutePath());
-        waitForEmptyConsumerGroup(adminClient, appID, TIMEOUT_MULTIPLIER * STREAMS_CONSUMER_TIMEOUT);
-
-        assertInternalTopicsGotDeleted(null);
-
-        resetFile.deleteOnExit();
-
-        // RE-RUN
-        streams.start();
-        final List<KeyValue<Long, Long>> resultRerun = IntegrationTestUtils.waitUntilMinKeyValueRecordsReceived(resultConsumerConfig, OUTPUT_TOPIC, 5);
-        streams.close();
-
-        result.remove(0);
-        assertThat(resultRerun, equalTo(result));
-
-        waitForEmptyConsumerGroup(adminClient, appID, TIMEOUT_MULTIPLIER * STREAMS_CONSUMER_TIMEOUT);
-        cleanGlobal(false, null, null);
-    }
-
-    @Test
-    public void testReprocessingFromDateTimeAfterResetWithoutIntermediateUserTopic() throws Exception {
-        appID = testId + "-from-datetime";
-        streamsConfig.put(StreamsConfig.APPLICATION_ID_CONFIG, appID);
-
-        // RUN
-        streams = new KafkaStreams(setupTopologyWithoutIntermediateUserTopic(), streamsConfig);
-        streams.start();
-        final List<KeyValue<Long, Long>> result = IntegrationTestUtils.waitUntilMinKeyValueRecordsReceived(resultConsumerConfig, OUTPUT_TOPIC, 10);
-
-        streams.close();
-        waitForEmptyConsumerGroup(adminClient, appID, TIMEOUT_MULTIPLIER * STREAMS_CONSUMER_TIMEOUT);
-
-        // RESET
-        final File resetFile = File.createTempFile("reset", ".csv");
-        try (final BufferedWriter writer = new BufferedWriter(new FileWriter(resetFile))) {
-            writer.write(INPUT_TOPIC + ",0,1");
-        }
-
-        streams = new KafkaStreams(setupTopologyWithoutIntermediateUserTopic(), streamsConfig);
-        streams.cleanUp();
-
-
-        final SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS");
-        final Calendar calendar = Calendar.getInstance();
-        calendar.add(Calendar.DATE, -1);
-
-        cleanGlobal(false, "--to-datetime", format.format(calendar.getTime()));
-        waitForEmptyConsumerGroup(adminClient, appID, TIMEOUT_MULTIPLIER * STREAMS_CONSUMER_TIMEOUT);
-
-        assertInternalTopicsGotDeleted(null);
-
-        resetFile.deleteOnExit();
-
-        // RE-RUN
-        streams.start();
-        final List<KeyValue<Long, Long>> resultRerun = IntegrationTestUtils.waitUntilMinKeyValueRecordsReceived(resultConsumerConfig, OUTPUT_TOPIC, 10);
-        streams.close();
-
-        assertThat(resultRerun, equalTo(result));
-
-        waitForEmptyConsumerGroup(adminClient, appID, TIMEOUT_MULTIPLIER * STREAMS_CONSUMER_TIMEOUT);
-        cleanGlobal(false, null, null);
-    }
-
-    @Test
-    public void testReprocessingByDurationAfterResetWithoutIntermediateUserTopic() throws Exception {
-        appID = testId + "-from-duration";
-        streamsConfig.put(StreamsConfig.APPLICATION_ID_CONFIG, appID);
-
-        // RUN
-        streams = new KafkaStreams(setupTopologyWithoutIntermediateUserTopic(), streamsConfig);
-        streams.start();
-        final List<KeyValue<Long, Long>> result = IntegrationTestUtils.waitUntilMinKeyValueRecordsReceived(resultConsumerConfig, OUTPUT_TOPIC, 10);
-
-        streams.close();
-        waitForEmptyConsumerGroup(adminClient, appID, TIMEOUT_MULTIPLIER * STREAMS_CONSUMER_TIMEOUT);
-
-        // RESET
-        final File resetFile = File.createTempFile("reset", ".csv");
-        try (final BufferedWriter writer = new BufferedWriter(new FileWriter(resetFile))) {
-            writer.write(INPUT_TOPIC + ",0,1");
-        }
-
-        streams = new KafkaStreams(setupTopologyWithoutIntermediateUserTopic(), streamsConfig);
-        streams.cleanUp();
-        cleanGlobal(false, "--by-duration", "PT1M");
-
-        waitForEmptyConsumerGroup(adminClient, appID, TIMEOUT_MULTIPLIER * STREAMS_CONSUMER_TIMEOUT);
-
-        assertInternalTopicsGotDeleted(null);
-
-        resetFile.deleteOnExit();
-
-        // RE-RUN
-        streams.start();
-        final List<KeyValue<Long, Long>> resultRerun = IntegrationTestUtils.waitUntilMinKeyValueRecordsReceived(resultConsumerConfig, OUTPUT_TOPIC, 10);
-        streams.close();
-
-        assertThat(resultRerun, equalTo(result));
-
-        waitForEmptyConsumerGroup(adminClient, appID, TIMEOUT_MULTIPLIER * STREAMS_CONSUMER_TIMEOUT);
-        cleanGlobal(false, null, null);
-    }
-
     private Topology setupTopologyWithIntermediateTopic(final boolean useRepartitioned,
                                                         final String outputTopic2) {
         final StreamsBuilder builder = new StreamsBuilder();
@@ -561,7 +339,7 @@ public abstract class AbstractResetIntegrationTest {
         return builder.build();
     }
 
-    private Topology setupTopologyWithoutIntermediateUserTopic() {
+    protected Topology setupTopologyWithoutIntermediateUserTopic() {
         final StreamsBuilder builder = new StreamsBuilder();
 
         final KStream<Long, String> input = builder.stream(INPUT_TOPIC);
@@ -573,7 +351,7 @@ public abstract class AbstractResetIntegrationTest {
         return builder.build();
     }
 
-    private boolean tryCleanGlobal(final boolean withIntermediateTopics,
+    protected boolean tryCleanGlobal(final boolean withIntermediateTopics,
                                    final String resetScenario,
                                    final String resetScenarioArg) throws Exception {
         // leaving --zookeeper arg here to ensure tool works if users add it
@@ -615,14 +393,14 @@ public abstract class AbstractResetIntegrationTest {
         return new StreamsResetter().run(parameters, cleanUpConfig) == 0;
     }
 
-    private void cleanGlobal(final boolean withIntermediateTopics,
+    protected void cleanGlobal(final boolean withIntermediateTopics,
                              final String resetScenario,
                              final String resetScenarioArg) throws Exception {
         final boolean cleanResult = tryCleanGlobal(withIntermediateTopics, resetScenario, resetScenarioArg);
         Assert.assertTrue(cleanResult);
     }
 
-    private void assertInternalTopicsGotDeleted(final String intermediateUserTopic) throws Exception {
+    protected void assertInternalTopicsGotDeleted(final String intermediateUserTopic) throws Exception {
         // do not use list topics request, but read from the embedded cluster's zookeeper path directly to confirm
         if (intermediateUserTopic != null) {
             cluster.waitForRemainingTopics(30000, INPUT_TOPIC, OUTPUT_TOPIC, OUTPUT_TOPIC_2, OUTPUT_TOPIC_2_RERUN,
